% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc_bathymetry}
\alias{bc_bathymetry}
\title{A list of length-3 lists containing points which can be used to generate contour lines on the BC coast}
\format{
A list
}
\usage{
bc_bathymetry
}
\description{
A list of length-3 lists containing points which can be used to generate contour lines on the BC coast
}
\keyword{datasets}
