% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{calc_landings}
\alias{calc_landings}
\title{Sum and count the landings by aggregating by years and months. Assumes the
catch weight is in pounds and converts the sum to kilograms.}
\usage{
calc_landings(df, fishery)
}
\arguments{
\item{df}{data frame as}

\item{fishery}{}
}
\value{
a data frame with the following five columns:
  Fishery, Year, Month, weightKg, numLandings
  Where Fishery will be the same for all records (fishery string)
}
\description{
Sum and count the landings by aggregating by years and months. Assumes the
catch weight is in pounds and converts the sum to kilograms.
}
