% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-age-data.R
\name{calc_lw_params}
\alias{calc_lw_params}
\title{Calculate the length-weight relationship parameters for the data with both length and weight, on a
specimen-by-specime basis}
\usage{
calc_lw_params(
  d,
  grouping_col = "year",
  lw_cutoff = 10,
  lw_tol = 0.1,
  lw_maxiter = 1000
)
}
\arguments{
\item{d}{The data frame as extracted using \code{\link[=fetch_sample_data]{fetch_sample_data()}}}

\item{grouping_col}{A character string matching the name of the column you want to group for}

\item{lw_cutoff}{How many length-weight records are required per sample to use the
length-weight model for that sample. If less than this, the overall yearly values will be used}

\item{lw_tol}{See \code{\link[=fit_lw]{fit_lw()}}}

\item{lw_maxiter}{See \code{\link[=fit_lw]{fit_lw()}}}
}
\value{
The original data frame with the columns \code{lw_alpha} and \code{lw_beta} added (if necessary). If those columns
exist, they will be overwritten with NAs and values calculated where data exist
}
\description{
Calculate the length-weight relationship parameters for the data with both length and weight, on a
specimen-by-specime basis
}
