% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{catch_by_day}
\alias{catch_by_day}
\title{catch_by_day}
\usage{
catch_by_day(d, major_areas = major_hake_areas,
  fishery = fishery_enum()$all, include_juandefuca = TRUE,
  byarea = FALSE)
}
\arguments{
\item{d}{a list of data retrieved using gfdata package functions}

\item{major_areas}{a vector of major stat areas as strings. e.g. "01" is 4B Strait of Georgia}

\item{fishery}{the fishery to return the catch for. Default is all records. Uses the fishery_enum
function as an enumerator to shorten names.}

\item{include_juandefuca}{Include the minor area of Juan De Fuca Strait which is located in major area 4B}

\item{byarea}{If TRUE, data frame will be have one row per unique date & area. If FALSE,
there will be one unique row for each day}
}
\value{
the catch data frame
}
\description{
Calculate the catch by day for a given fishery
}
\examples{
fetch_data()
d <- load_data()
ct <- catch_by_day(d)
ct.ft <- catch_by_day(d, fishery = fishery_enum()$ft)
}
