% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coast_10m}
\alias{coast_10m}
\title{An sf LINESTRING object which contains data for 10m resolution for the world coastline. Data source:
#' http://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/physical/ne_10m_coastline.zip}
\format{
An sf LINESTRING object
}
\usage{
coast_10m
}
\description{
An sf LINESTRING object which contains data for 10m resolution for the world coastline. Data source:
#' http://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/physical/ne_10m_coastline.zip
}
\keyword{datasets}
