% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-spatial-catch-data.R
\name{conv_spatial}
\alias{conv_spatial}
\title{Merge all dataframes into one spatial dataframe. See \link{sf} package}
\usage{
conv_spatial(...)
}
\arguments{
\item{...}{Dataframes to merge. They must all have the columns \code{X} and \code{Y}}
}
\value{
a spatial data frame as described in the \link{sf} package
}
\description{
This function converts all supplied data frames into a single \link{sf} data frame. It also
removes points found on land or way out of the fishing area. The Fishing IDs (FIDs) for
this filtration were found using \code{\link[mapview:mapview]{mapview::mapview()}}.
}
