% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{create_catch_csv_file}
\alias{create_catch_csv_file}
\title{Create a CSV file with catch by year and month for a particular fishery}
\usage{
create_catch_csv_file(df, fn = NA)
}
\arguments{
\item{df}{The data frame to reduce into a CSV. This will be output from the \code{\link[=get_catch]{get_catch()}} function}

\item{fn}{Filename for the CSV}
}
\description{
Create a CSV file with catch by year and month for a particular fishery
}
