% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmp_catch_sql_file}
\alias{dmp_catch_sql_file}
\title{Name of the DMP catch SQL file}
\format{
A character string
}
\usage{
dmp_catch_sql_file
}
\description{
Name of the DMP catch SQL file
}
\keyword{datasets}
