% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmp_file}
\alias{dmp_file}
\title{Name of the Dockside Monitoring Program data file (CSV file)}
\format{A character string}
\usage{
dmp_file
}
\description{
Name of the Dockside Monitoring Program data file (CSV file)
}
\keyword{datasets}
