% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{.onLoad}
\alias{.onLoad}
\title{When the package is loaded, set ggplot theme to hake theme}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{See \code{\link[=.onLoad]{.onLoad()}}}

\item{pkgname}{See \code{\link[=.onLoad]{.onLoad()}}}
}
\description{
When the package is loaded, set ggplot theme to hake theme
}
