% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{far_point}
\alias{far_point}
\title{An sf class POINT which represents a point at the lower rightmost part of the fishery distribution
to be used to measure disctance to other point for removal of outlier points}
\format{
An sf POINT object
}
\usage{
far_point
}
\description{
An sf class POINT which represents a point at the lower rightmost part of the fishery distribution
to be used to measure disctance to other point for removal of outlier points
}
\keyword{datasets}
