% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{fetch_catch}
\alias{fetch_catch}
\title{Fetch the DMP catch from the FOS database}
\usage{
fetch_catch(end_date = format(Sys.Date(), "\%d/\%m/\%Y"), overwrite = FALSE)
}
\arguments{
\item{end_date}{The end date of the date extraction}

\item{overwrite}{Logical. Overwrite the RDS files for DMP and LOGS data if they exist}
}
\value{
Nothing
}
\description{
Fetch the DMP catch from the FOS database
}
