% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-catch-data.R
\name{fetch_catch_data}
\alias{fetch_catch_data}
\title{Fetch the DMP catch from the FOS database}
\usage{
fetch_catch_data(
  end_date = format(Sys.Date(), "\%d/\%m/\%Y"),
  overwrite = FALSE
)
}
\arguments{
\item{end_date}{The end date of the date extraction}

\item{overwrite}{Logical. Overwrite the RDS files for DMP and LOGS data if they exist}
}
\value{
Nothing
}
\description{
Fetch the DMP catch from the FOS database
}
