% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{fetch_data}
\alias{fetch_data}
\title{fetch_data}
\usage{
fetch_data(file = here("generated-data", paste0(gsub(" ", "-",
  species_name), ".rds")))
}
\arguments{
\item{file}{the full path filename including extension .rds}
}
\description{
fetch_data
}
\examples{
fetch_data()
}
