% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-age-data.R
\name{fetch_sample_data}
\alias{fetch_sample_data}
\title{Fetch the sample data from the GFBIOSQL database
The data will be filtered to only include hake major and minor areas}
\usage{
fetch_sample_data(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Logical. Overwrite the RDS file for sample data if it exists}
}
\description{
Fetch the sample data from the GFBIOSQL database
The data will be filtered to only include hake major and minor areas
}
