% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-spatial-catch-data.R
\name{fetch_spatial_catch_data}
\alias{fetch_spatial_catch_data}
\title{Run SQL code on the server database and return the records}
\usage{
fetch_spatial_catch_data(type, overwrite_file = FALSE)
}
\arguments{
\item{type}{One of 'ft', 'ss', or 'jv' for Freezer Trawler, Shoreside, and Joint Venture respectively}

\item{overwrite_file}{If TRUE, and the function has been run before and an rds file generated,
overwrite this. If FALSE, run the SQL and generate the file.}
}
\value{
A data frame of the records returned for the SQL code given
}
\description{
Run SQL code on the server database and return the records
}
