% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{fishery_enum}
\alias{fishery_enum}
\title{fishery_enum}
\usage{
fishery_enum()
}
\value{
A character astring representing the fishery
}
\description{
An enumeration function used to generalize code used in other functions
}
\examples{
fishery_enum()$ft
}
