% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_alw}
\alias{get_alw}
\title{get_alw}
\usage{
get_alw(d)
}
\arguments{
\item{d}{a list of data retrieved using gfplot package functions}
}
\value{
the age/sex/length by year data frame
}
\description{
get_alw
}
