% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_catch}
\alias{get_catch}
\title{Get the catch data from the DMP and LOGS data frames}
\usage{
get_catch(dmp, logs, type)
}
\arguments{
\item{dmp}{The Dockside Monitoring Program data frame as extracted by [hakedata::load_data()]}

\item{logs}{The Logbook data frame as extracted by [hakedata::load_data()]}

\item{type}{one of "ft", "ss", or "jv" for Freezer Trawler, Shoreside, and Joint Venture respectively}
}
\value{
a data frame containing only freezer trawler data
}
\description{
Get the catch data from the DMP and LOGS data frames
}
