% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_comm_samples}
\alias{get_comm_samples}
\title{Fetch commercial hake data from the database or local file}
\usage{
get_comm_samples(rerun_query = FALSE, major_areas = major_hake_areas,
  include_juandefuca = TRUE)
}
\arguments{
\item{rerun_query}{Run the query to fetch data, and re-write the RDS file for future use.}

\item{major_areas}{The major statistical areas to include in the data}

\item{include_juandefuca}{Include the Strait of Juan De Fuca in the data? TREU/FALSE}
}
\value{
A data frame of the all the commercial samples for hake found in the areas given
}
\description{
Fetch commercial hake data from the database or local file
}
