% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wa.R
\name{get_comm_wa}
\alias{get_comm_wa}
\title{Get the weight-at-age data frame for commercial data.}
\usage{
get_comm_wa(d)
}
\arguments{
\item{d}{}
}
\value{
a data frame with Source set to \code{CAN_shoreside} and \code{CAN_freezer}
for the two fishery types
}
\description{
Get the weight-at-age data frame for commercial data.
}
