% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-depths.R
\name{get_depth_by_year}
\alias{get_depth_by_year}
\title{Get depth data by year and calculate boxplot stats on it.
This writes a CSV file containing the output}
\usage{
get_depth_by_year(d, type = "bottom", yrs = NULL, min_depth_cutoff = 50/1.8288)
}
\arguments{
\item{d}{Output from \code{\link[=load_spatial_catch_data]{load_spatial_catch_data()}}}

\item{type}{One of "bottom" or "gear" for depth type}

\item{yrs}{A vector of years to include. If NULL, all years in the data will be included}

\item{min_depth_cutoff}{The depth for which to remove data. In fasthoms. All data shallower than this
will be removed. This applies to both \code{type} 'bottom' and 'gear'}
}
\value{
Invisibly - A tibble containing year and depth record stats
}
\description{
Get depth data by year and calculate boxplot stats on it.
This writes a CSV file containing the output
}
\examples{
d_ss <- load_spatial_catch_data("ss")
gear_depth_ss <- get_depth_by_year(d_ss, "gear")
}
