% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_ft_catch}
\alias{get_ft_catch}
\title{Get the freezer trawler data from the DMP and LOGS data frames}
\usage{
get_ft_catch(dmp, logs)
}
\arguments{
\item{dmp}{The Dockside Monitoring Program data frame as extracted by [hakedata::load_data()]}

\item{logs}{The Logbook data frame as extracted by [hakedata::load_data()]}
}
\value{
a data frame containing only freezer trawler data
}
\description{
Get the freezer trawler data from the DMP and LOGS data frames
}
