% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wa.R
\name{get_surv_wa}
\alias{get_surv_wa}
\title{Get the weight-at-age data frame for the acoustic survey}
\usage{
get_surv_wa(d)
}
\arguments{
\item{d}{}
}
\value{
a data frame
}
\description{
Get the weight-at-age data frame for the acoustic survey
}
