% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hake_catch_jv_file}
\alias{hake_catch_jv_file}
\title{Name of the joint venture catch rds file}
\format{A character string}
\usage{
hake_catch_jv_file
}
\description{
Name of the joint venture catch rds file
}
\keyword{datasets}
