% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hakedata-package.R
\docType{package}
\name{hakedata}
\alias{hakedata}
\alias{hakedata-package}
\title{hakedata package - Extract Pacific hake data and plot it}
\description{
hakedata package - Extract Pacific hake data and plot it
}
