% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbs_to_kilos}
\alias{lbs_to_kilos}
\title{Conversion factor from pounds to kilograms}
\format{
A double
}
\usage{
lbs_to_kilos
}
\description{
Conversion factor from pounds to kilograms
}
\keyword{datasets}
