% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{load_data}
\alias{load_data}
\title{Loads data from csv files found in the data directory}
\usage{
load_data()
}
\value{
A list of three dataframes, \verb{[[1]]} for the DMP data, \verb{[[2]]} for the LOGS data, and \verb{[[3]]} for the area 4B
data which were not included. \verb{[[3]]} is returned as a convinience for you to see if any other data from area 4B
should be included
}
\description{
See \link{dmp_file} and \link{logs_pattern} for descriptions of filenames
}
