% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{load_data_gfdata}
\alias{load_data_gfdata}
\title{load_data_gfdata}
\usage{
load_data_gfdata(file = here("generated-data", paste0(gsub(" ", "-",
  species_name), ".rds")))
}
\arguments{
\item{file}{the full path filename including extension .rds}
}
\value{
the contents of the rds file as a list
}
\description{
load_data_gfdata
}
\examples{
d <- load_data_gfdata()
}
