% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sample-data.R
\name{load_sample_data}
\alias{load_sample_data}
\title{load the commercial sample data from the file generated by \code{\link[=fetch_sample_data_gfdata]{fetch_sample_data_gfdata()}}}
\usage{
load_sample_data()
}
\value{

}
\description{
load the commercial sample data from the file generated by \code{\link[=fetch_sample_data_gfdata]{fetch_sample_data_gfdata()}}
}
