% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-catch-data.R
\name{load_spatial_catch_data}
\alias{load_spatial_catch_data}
\title{Loads spatial catch data from RDS files found in the data directory}
\usage{
load_spatial_catch_data(fishery = "ft")
}
\value{
The tibble with the data for the fishery given
}
\description{
Loads spatial catch data from RDS files found in the data directory
}
