% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-survey-data.R
\name{load_survey_data}
\alias{load_survey_data}
\title{Loads survey data from RDS files found in the data directory}
\usage{
load_survey_data()
}
\value{
The tibble with the data
}
\description{
Loads survey data from RDS files found in the data directory
}
