% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logs_catch_data_raw_file}
\alias{logs_catch_data_raw_file}
\title{Name of the LOGS catch rds file}
\format{A character string}
\usage{
logs_catch_data_raw_file
}
\description{
Name of the LOGS catch rds file
}
\keyword{datasets}
