% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logs_pattern}
\alias{logs_pattern}
\title{Pattern for regular expressions to use for LOGS data files}
\format{
A character string
}
\usage{
logs_pattern
}
\description{
Pattern for regular expressions to use for LOGS data files
}
\keyword{datasets}
