% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{major_hake_areas}
\alias{major_hake_areas}
\title{Major PFMA areas used when assessing hake}
\format{
A character vector
}
\usage{
major_hake_areas
}
\description{
Major PFMA areas used when assessing hake
}
\keyword{datasets}
