% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_area_dist}
\alias{plot_area_dist}
\title{Plot distribution of catch by area}
\usage{
plot_area_dist(d, inc_areas = major_hake_areas,
  inc_years = (year(now()) - 5):year(now()), ylim = c(0, 1300))
}
\arguments{
\item{d}{Data frame as output from catch_by_day()}

\item{inc_areas}{Areas to include in the plot}

\item{inc_years}{Years to include in the plot}

\item{ylim}{Vector of 2, Y-axis limits}
}
\value{
A ggplot object
}
\description{
Plot distribution of catch by area
}
\examples{
d <- load_data()
ct <- catch_by_day(d, byarea = TRUE)
ct.ft <- catch_by_day(d, fishery = fishery_enum()$ft, byarea = TRUE)
plot_area_dist(ct.ft, ylim = c(0, 45))
plot_area_dist(ct)
}
