% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_color}
\alias{plot_color}
\title{Colors for the Hake Assessment}
\usage{
plot_color(n.cols = 10)
}
\arguments{
\item{n}{The default number of colours to generate is 10.}
}
\value{
A vector of colours
}
\description{
Generate the color scheme for the hake assessment
such that as more years of historical data are added on
that the more recent years are still the same color.
For example, the most recent year of data will always be
black and the second most recent year will always be red.
}
\examples{
n <-18
plot(data.frame(1:n, 1), col = plot_color(n), pch = 19, cex = 5)
}
\author{
Kelli Faye Johnson
}
