% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_cumu_catch}
\alias{plot_cumu_catch}
\title{Plot_cumu_catch}
\usage{
plot_cumu_catch(
  d,
  inc_years = (year(now()) - 5):year(now()),
  weight_factor = 1e+06,
  ylim = c(0, 100),
  line_thickness = 1.5,
  horiz_line_spacing = NA
)
}
\arguments{
\item{d}{Data frame as output from \code{\link[=get_catch]{get_catch()}}}

\item{inc_years}{Years to include in the plot}

\item{weight_factor}{Divide catch values by this to scale y-axis}

\item{ylim}{Vector of 2, Y-axis limits}

\item{line_thickness}{thickness of the lines}

\item{horiz_line_spacing}{spacing to draw horizontal dotted lines at. If NA,
no lines will be plotted}
}
\value{
A ggplot object
}
\description{
Plot the cumulative catch for the given years
}
