% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_spatial}
\alias{plot_spatial}
\title{Spatial plot of data by block}
\usage{
plot_spatial(
  grd,
  data_col = "num_fids",
  data_factor = 1,
  crs = 4326,
  extents = data.frame(lon = c(-135, -122), lat = c(48, 55)),
  contour_depths = c(100, 200, 400, 1000, 1500, 2000),
  contour_color = "lightblue",
  contour_thickness = 0.25
)
}
\arguments{
\item{grd}{A list of length 2, the output of the \code{\link[=make_grid]{make_grid()}} function}

\item{data_col}{The column in the \code{grd} data frame to use as the plotting value}

\item{data_factor}{Value to divide the \code{data_col} values by in the \code{grd} data frame}

\item{crs}{See \code{\link[=contours_as_sfg]{contours_as_sfg()}}}

\item{extents}{A data frame with two columns, named 'lon' and 'lat' which represent the
extents of the plotting area. The data frame must have two rows and two columns}

\item{contour_depths}{A vector of depths to plot. Must be already present in the contour data
(\link{bc_bathymetry}). If NA, no contours will be plotted}

\item{contour_color}{Color for the contour lines}

\item{contour_thickness}{Thickness of contour lines}
}
\value{
A ggplot object
}
\description{
Spatial plot of data by block
}
