% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{read_sql}
\alias{read_sql}
\title{Read in SQL code from a package file}
\usage{
read_sql(fn)
}
\arguments{
\item{fn}{Filename for SQL code}
}
\value{
a vector of charcter strings, one for each line in `fn``
}
\description{
Read in SQL code from a package file
}
