% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_raw_file}
\alias{sample_data_raw_file}
\title{Name of the commerical samples rds file}
\format{
A character string
}
\usage{
sample_data_raw_file
}
\description{
Name of the commerical samples rds file
}
\keyword{datasets}
