% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_catch_sql_file}
\alias{spatial_catch_sql_file}
\title{Name of the spatial catch SQL template file}
\format{A character string}
\usage{
spatial_catch_sql_file
}
\description{
Name of the spatial catch SQL template file
}
\keyword{datasets}
