% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_name}
\alias{species_name}
\title{Constants for the hakedata package}
\format{An object of class \code{character} of length 1.}
\usage{
species_name
}
\description{
species code, data frame of freezer trawler vessel codes,
and hake major areas are included
}
\keyword{datasets}
