% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{strip_lines}
\alias{strip_lines}
\title{Strip the first n lines off a file, and re-save}
\usage{
strip_lines(file)
}
\arguments{
\item{file}{filename}
}
\description{
Strip the first n lines off a file, and re-save
}
