% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data_file}
\alias{survey_data_file}
\title{Name of the survey data rds file}
\format{
A character string
}
\usage{
survey_data_file
}
\description{
Name of the survey data rds file
}
\keyword{datasets}
