% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-age-data.R
\name{get_age_props}
\alias{get_age_props}
\title{Calculate the age proportions}
\usage{
get_age_props(
  min_date = as.Date("1985-01-01"),
  plus_grp = 15,
  lw_model_by_year = FALSE
)
}
\arguments{
\item{min_date}{Earliest date to include}

\item{plus_grp}{Age plus group for maximum grouping}
}
\value{

}
\description{
Calculate the age proportions
}
