% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_catch}
\alias{get_catch}
\title{Get the catch data from the DMP and LOGS data frames}
\usage{
get_catch(d, type)
}
\arguments{
\item{d}{a list of length 2 with elements being (1) The Dockside Monitoring Program
data frame as extracted by \code{\link[hakedata:load_data]{hakedata::load_data()}} and (2) the Logbook data frame
as extracted by \code{\link[hakedata:load_data]{hakedata::load_data()}}}

\item{type}{one of "ft", "ss", or "jv" for Freezer Trawler, Shoreside, and Joint Venture respectively}
}
\value{
a data frame containing only freezer trawler data
}
\description{
Get the catch data from the DMP and LOGS data frames
}
