% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-catch-data.R
\name{get_catch}
\alias{get_catch}
\title{Get the catch data from the DMP and LOGS data frames}
\usage{
get_catch(d, type)
}
\arguments{
\item{d}{a list of length 2 with elements being (1) The Dockside Monitoring Program
data frame as extracted by \code{\link[=load_catch_data]{load_catch_data()}} and (2) the Logbook data frame
as extracted by \code{\link[=load_catch_data]{load_catch_data()}}}

\item{type}{one of "ft", "ss", or "jv" for Freezer Trawler, Shoreside, or Joint Venture respectively}
}
\value{
A data frame containing only data for the type given by \code{type}
}
\description{
Get the catch data from the DMP and LOGS data frames
}
