% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_spatial_catch_sql}
\alias{get_spatial_catch_sql}
\title{Run SQL code on the server database and return the records}
\usage{
get_spatial_catch_sql(type)
}
\arguments{
\item{type}{one of "ft", "ss", or "jv" for Freezer Trawler, Shoreside, and Joint Venture respectively}
}
\value{
a data frame of the records for the SQL code given
}
\description{
Run SQL code on the server database and return the records
}
