% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hakedata-package.R
\docType{package}
\name{hakedata}
\alias{hakedata}
\alias{hakedata-package}
\title{hakedata package}
\description{
Extract Pacific hake data and plot it. Follow the instructions here:
\url{https://github.com/pbs-assess/hake} to extract data, and see the rmarkdown file:
\url{https://github.com/pbs-assess/hake/blob/master/hake-data.Rmd} for an example.
}
