% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{inject_fishery_filter}
\alias{inject_fishery_filter}
\title{Inject SQL code for fihery categories and types based on the fleet type for hake}
\usage{
inject_fishery_filter(sql, type)
}
\arguments{
\item{sql}{SQL code as a vector of character strings (from \code{\link[base:readLines]{base::readLines()}})}

\item{type}{one of "ft", "ss", or "jv" for Freezer Trawler, Shoreside, and Joint Venture respectively}
}
\description{
Inject SQL code for fihery categories and types based on the fleet type for hake
}
