% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{load_data}
\alias{load_data}
\title{Loads data from csv files found in the data directory. See [hakedata::landings_file]
and [hakedata::logs_pattern] for filenames.}
\usage{
load_data()
}
\value{
a list of two dataframes, one for the DMP data and one for the LOGS data
}
\description{
Loads data from csv files found in the data directory. See [hakedata::landings_file]
and [hakedata::logs_pattern] for filenames.
}
