% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(file = file.path(here::here("generated-data"), paste0(gsub(" ",
  "-", hakedata::species_name), ".rds")))
}
\arguments{
\item{file}{the full path filename including extension .rds}
}
\value{
the contents of the rds file as a list
}
\description{
load_data
}
\examples{
d <- load_data()
}
