% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_grid}
\alias{make_grid}
\title{Create a grid based on fishing event data.}
\usage{
make_grid(d, cell_size = 10000, min_num_fids = 3)
}
\arguments{
\item{d}{An \code{sf} \code{POINT} object}

\item{cell_size}{Cell side length in meters}

\item{min_num_fids}{Minimum number of FIDs in a cell. If less than this, they will be set to zero}
}
\value{
An \code{sf} \code{MULTIPOLYGON} object with columns added: \code{cell} and \code{num_fids}
}
\description{
Creates a grid of cells for which fishing event data exist. The cells with zero fishing
events and cells which have less than \code{min_num_fids} will be removed.
}
