% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{merge_into_spatial}
\alias{merge_into_spatial}
\title{Merge all dataframes into one spatial dataframe. See [sf] package.}
\usage{
merge_into_spatial(..., crs = 4326)
}
\arguments{
\item{...}{dataframes to merge. They must all have the columns `X` and `Y`}

\item{crs}{see [sf::st_sf()]}
}
\value{
a spatial data frame as described in the [sf] package
}
\description{
Merge all dataframes into one spatial dataframe. See [sf] package.
}
