% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_cumu_catch}
\alias{plot_cumu_catch}
\title{Plot_cumu_catch}
\usage{
plot_cumu_catch(d, inc_years = 2009:lubridate::year(lubridate::now()),
  weight_factor = 1e+06, ylim = c(0, 45), horiz.line.spacing = NA)
}
\arguments{
\item{d}{Data frame as output from catch_by_day()}

\item{inc_years}{Years to include in the plot}

\item{weight_factor}{Divide catch values by this to scale y-axis}

\item{ylim}{Vector of 2, Y-axis limits}

\item{horiz.line.spacing}{}
}
\value{
A ggplot object
}
\description{
Plot the cumulative catch for the given years
}
\examples{
d <- load_data()
ct <- catch_by_day(d)
ct.ft <- catch_by_day(d, fishery = fishery_enum()$ft)
plot_cumu_catch(ct.ft, horiz.line.spacing = 5)
}
