% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_depths}
\alias{plot_depths}
\title{Plot depth boxplots by year (catchdate) or histograms for a single year}
\usage{
plot_depths(d, plot_type = "hist", fishery_type = NA, bin_width = 10,
  max_depth_shown = 600, alpha = 0.6, legend_loc = "inside")
}
\arguments{
\item{d}{The data frame as returned by the spatial SQL queries found in this in
the variable \code{spatial_catch_sql_file}}

\item{plot_type}{Type of plot, one of 'hist' or 'box'}

\item{fishery_type}{One of NA, 'ft', 'ss', or 'jv'. If NA, all fisheries will be plotted against one another,
if one of the others, that will be filtered before plotting}

\item{bin_width}{See \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}}

\item{max_depth_shown}{Maximum depth to be shown. All records with a greater depth are filtered out
prior to plotting using \code{\link[dplyr:filter]{dplyr::filter()}}}

\item{alpha}{Transparency of fill from 0 - 1}

\item{legend_loc}{Where to place legend "inside" or "outside" the frame of the plot}
}
\value{
A \link{ggplot2} object
}
\description{
Plot depth boxplots by year (catchdate) or histograms for a single year
}
