% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_lengths}
\alias{plot_lengths}
\title{Plot length distributions as histograms}
\usage{
plot_lengths(yrs = NA, bin_width = 5, xlim = c(NA, NA),
  alpha = 0.6, legend_loc = "inside")
}
\arguments{
\item{yrs}{Which years to include in the data to plot. If NA, all years in the \code{Years} column
will be used}

\item{bin_width}{See \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}}

\item{xlim}{Limits for the x-axis. For c(NA, NA), plot will extend to limits of data}

\item{alpha}{Transparency of fill from 0 - 1}

\item{legend_loc}{Where to place legend "inside" or "outside" the frame of the plot}
}
\value{
A \link{ggplot2} object
}
\description{
Plot length distributions as histograms
}
