% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{remove_trip_data}
\alias{remove_trip_data}
\title{Remove some trip types from a data frame}
\usage{
remove_trip_data(d, types, trip_type_colname)
}
\arguments{
\item{d}{A data frame containing the column \code{trip_type_colname}}

\item{types}{A vector of strings to grep for in the \code{trip_type_colname} column for removal}

\item{trip_type_colname}{The column name containing the trip type strings}
}
\value{
A data frame with the data removed for the trip types
}
\description{
Remove some trip types from a data frame
}
