% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-age-data.R
\name{sample_summary}
\alias{sample_summary}
\title{Summarize the numbers of lengths, weights, ages, sample_weights, and catch_weights in the sample data}
\usage{
sample_summary(d = readRDS(here("data", sample_data_raw_file)))
}
\arguments{
\item{d}{A dataframe as extracted using \code{\link[=fetch_sample_data]{fetch_sample_data()}}}
}
\value{
A list of two dataframes, the first summarized by year and the second sampled by sample_id and year
}
\description{
Summarize the numbers of lengths, weights, ages, sample_weights, and catch_weights in the sample data
}
