% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_name}
\alias{species_name}
\title{Constants for the hakedata package}
\format{A character string}
\usage{
species_name
}
\description{
A dataset containing the major statistical area codes and descriptions.
}
\details{
Name of the species
}
\keyword{datasets}
