% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-catch-data.R
\name{strip_lines}
\alias{strip_lines}
\title{Strip header lines of LOGBOOK file}
\usage{
strip_lines(file)
}
\arguments{
\item{file}{filename (csv file)}
}
\description{
Search the file for the first occurance of the line starting with 'LOGBOOK'
and remove it, and any following blank lines leaving only the column headers and data.
Overwrite the file with this.
}
